+++ scripts/build_renewable_profiles.py
@@ -192,6 +192,7 @@ import functools
 import logging
 import os
 import time
+from math import isnan
 
 import atlite
 import country_converter as coco
@@ -203,7 +204,7 @@ import xarray as xr
 from _helpers import configure_logging, read_csv_nafix, sets_path_to_root
 from add_electricity import load_powerplants
 from pypsa.geo import haversine
-from shapely.geometry import LineString
+from shapely.geometry import LineString, Point
 
 cc = coco.CountryConverter()
 
@@ -227,13 +228,12 @@ def get_eia_annual_hydro_generation(fn, countries):
     df.index.name = "countries"
 
     df = df.T[countries] * 1e6  # in MWh/a
+    df.index = df.index.astype(int)
 
     return df
 
 
-def get_hydro_capacity_annual_hydro_generation(
-    fn, countries, year_start=2000, year_end=2020
-):
+def get_hydro_capacity_annual_hydro_generation(fn, countries, year):
     hydro_stats = (
         pd.read_csv(
             fn,
@@ -252,15 +252,9 @@ def get_hydro_capacity_annual_hydro_generation(
         * 1e3
         * 8760
     )  # change unit to MWh/y
+    hydro_prod_by_country.index = pd.Index([year])
 
-    range_years = range(year_start, year_end + 1)
-
-    normalize_using_yearly = pd.concat(
-        [hydro_prod_by_country] * len(range_years), ignore_index=True
-    )
-    normalize_using_yearly.index = pd.Index(range_years)
-
-    return normalize_using_yearly
+    return hydro_prod_by_country
 
 
 def check_cutout_completness(cf):
@@ -312,12 +306,154 @@ def filter_cutout_region(cutout, regions):
     return cutout
 
 
+def normalize_hydro(plants, runoff, normalize_using_yearly, normalization_year):
+    """
+    Function used to normalize the inflows of the hydro capacities to match country statistics
+
+    Parameters
+    ----------
+    plants : DataFrame
+        Run-of-river plants or dams with lon, lat, countries, installed_hydro columns.
+        Countries and installed_hydro column are only used with normalize_using_yearly
+        installed_hydro column shall be a boolean vector specifying whether that plant
+        is currently installed and used to normalize the inflows
+    runoff : xarray object
+        Runoff at each bus
+    normalize_using_yearly : DataFrame
+        Dataframe that specifies for every country the total hydro production
+    year : int
+        Year used for normalization
+    """
+
+    if plants.empty or plants.installed_hydro.any() == False:
+        logger.info("No bus has installed hydro plants, ignoring normalization.")
+        return runoff
+
+    years_statistics = normalize_using_yearly.index
+    if isinstance(years_statistics, pd.DatetimeIndex):
+        years_statistics = years_statistics.year
+    else:
+        years_statistics = years_statistics.astype(int)
+
+    years_statistics = years_statistics.unique()
+    years_runoff = pd.to_datetime(runoff.time).year.unique()
+
+    if normalization_year not in set(years_statistics):
+        logger.warning(
+            f"Missing hydro statistics for year {normalization_year}; no normalization performed."
+        )
+    else:
+        # get buses that have installed hydro capacity to be used to compute
+        # the normalization
+        normalization_buses = plants[plants.installed_hydro == True].index
+
+        # average yearly runoff by plant
+        yearlyavg_runoff_by_plant = (
+            runoff.rename("runoff").mean("time").to_dataframe()
+        ) * 8760.0
+
+        yearlyavg_runoff_by_plant["country"] = plants.loc[
+            yearlyavg_runoff_by_plant.index, "countries"
+        ]
+
+        # group runoff by country
+        grouped_runoffs = (
+            yearlyavg_runoff_by_plant.loc[normalization_buses].groupby("country").sum()
+        )
+
+        # common country indeces
+        common_countries = normalize_using_yearly.columns.intersection(
+            grouped_runoffs.index
+        )
+
+        tot_common_yearly = np.nansum(
+            normalize_using_yearly.loc[normalization_year, common_countries]
+        )
+        tot_common_runoff = np.nansum(grouped_runoffs.runoff[common_countries])
+
+        # define default_factor. When nan values, used the default 1.0
+        default_factor = 1.0
+        if not (
+            isnan(tot_common_yearly)
+            or isnan(tot_common_runoff)
+            or tot_common_runoff <= 0.0
+        ):
+            default_factor = tot_common_yearly / tot_common_runoff
+
+        def create_scaling_factor(
+            normalize_yearly, grouped_runoffs, year, c_bus, default_value=1.0
+        ):
+            if c_bus in normalize_yearly.columns and c_bus in grouped_runoffs.index:
+                # normalization in place
+                return (
+                    np.nansum(normalize_yearly.loc[year, c_bus])
+                    / grouped_runoffs.runoff[c_bus]
+                )
+            elif c_bus not in normalize_yearly.columns:
+                # data not available in the normalization procedure
+                # return unity factor
+                return default_value
+            elif c_bus not in grouped_runoffs.index:
+                # no hydro inflows available for he country
+                return default_value
+
+        unique_countries = plants.countries.unique()
+        missing_countries_normalization = np.setdiff1d(
+            unique_countries, normalize_using_yearly.columns
+        )
+        missing_countries_grouped_runoff = np.setdiff1d(
+            unique_countries, grouped_runoffs.index
+        )
+
+        if missing_countries_normalization.size != 0:
+            logger.warning(
+                f"Missing countries in the normalization dataframe: "
+                + ", ".join(missing_countries_normalization)
+                + ". Default value used"
+            )
+
+        if missing_countries_grouped_runoff.size != 0:
+            logger.warning(
+                f"Missing installed plants in: "
+                + ", ".join(missing_countries_grouped_runoff)
+                + ". Default value used"
+            )
+
+        # matrix used to scale the runoffs
+        scaling_matrix = xr.DataArray(
+            [
+                create_scaling_factor(
+                    normalize_using_yearly,
+                    grouped_runoffs,
+                    normalization_year,
+                    c_bus,
+                    default_factor,
+                )
+                * np.ones(runoff.time.shape)
+                for c_bus in plants.countries
+            ],
+            coords=dict(
+                plant=plants.index.values,
+                time=runoff.time.values,
+            ),
+        )
+
+        # Check all buses to be in the final dataset
+        missing_buses = plants.index.difference(scaling_matrix.plant)
+        if len(missing_buses) > 0:
+            logger.warning(f"Missing hydro inflows for buse {missing_buses}")
+
+        runoff *= scaling_matrix
+
+    return runoff
+
+
 if __name__ == "__main__":
     if "snakemake" not in globals():
         from _helpers import mock_snakemake
 
         os.chdir(os.path.dirname(os.path.abspath(__file__)))
-        snakemake = mock_snakemake("build_renewable_profiles", technology="onwind")
+        snakemake = mock_snakemake("build_renewable_profiles", technology="hydro")
         sets_path_to_root("pypsa-earth")
     configure_logging(snakemake)
 
@@ -396,60 +532,42 @@ if __name__ == "__main__":
         # if extendable option is true, all buses are included
         # otherwise only where hydro powerplants are available are considered
         if snakemake.config["cluster_options"]["alternative_clustering"]:
-            busbus_to_consider = [
-                (
-                    config.get("extendable", False)
-                    | (bus_id in hydro_ppls.region_id.values)
-                )
-                & any(hydrobasins.geometry.intersects(p))
-                for (p, bus_id) in zip(
-                    gpd.points_from_xy(regions.x, regions.y, crs=regions.crs),
-                    regions.shape_id,
-                )
-            ]
+            filter_bus_to_consider = regions.index.map(
+                lambda bus_id: config.get("extendable", False)
+                | (bus_id in hydro_ppls.region_id.values)
+            )
         ### TODO: quickfix. above case and the below case should by unified
         if snakemake.config["cluster_options"]["alternative_clustering"] == False:
-            busbus_to_consider = [
-                (config.get("extendable", False) | (bus_id in hydro_ppls.bus.values))
-                & any(hydrobasins.geometry.intersects(p))
-                for (p, bus_id) in zip(
-                    gpd.points_from_xy(regions.x, regions.y, crs=regions.crs),
-                    regions.index,
-                )
-            ]
+            filter_bus_to_consider = regions.index.map(
+                lambda bus_id: config.get("extendable", False)
+                | (bus_id in hydro_ppls.bus.values)
+            )
+        bus_to_consider = regions.index[filter_bus_to_consider]
+
+        # identify subset of buses within the hydrobasins
+        filter_bus_in_hydrobasins = regions[filter_bus_to_consider].apply(
+            lambda row: any(hydrobasins.geometry.contains(Point(row["x"], row["y"]))),
+            axis=1,
+        )
+        bus_in_hydrobasins = filter_bus_in_hydrobasins[filter_bus_in_hydrobasins].index
+
+        bus_notin_hydrobasins = list(
+            set(bus_to_consider).difference(set(bus_in_hydrobasins))
+        )
 
         resource["plants"] = regions.rename(
             columns={"x": "lon", "y": "lat", "country": "countries"}
-        ).loc[busbus_to_consider, ["lon", "lat", "countries"]]
+        ).loc[bus_in_hydrobasins, ["lon", "lat", "countries"]]
 
         resource["plants"]["installed_hydro"] = [
             True if (bus_id in hydro_ppls.bus.values) else False
             for bus_id in resource["plants"].index
         ]
 
-        # check if normalization field belongs to the settings and it is not false
-        if ("normalization" in resource) & (type(resource["normalization"]) == str):
-            normalization = resource.pop("normalization")
-
-            if normalization == "hydro_capacities":
-                path_hydro_capacities = snakemake.input.hydro_capacities
-                normalize_using_yearly = get_hydro_capacity_annual_hydro_generation(
-                    path_hydro_capacities, countries
-                ) * config.get("normalization_multiplier", 1.0)
-
-            elif normalization == "eia":
-                path_eia_stats = snakemake.input.eia_hydro_generation
-                normalize_using_yearly = get_eia_annual_hydro_generation(
-                    path_eia_stats, countries
-                ) * config.get("normalization_multiplier", 1.0)
-
-            resource["normalize_using_yearly"] = normalize_using_yearly
-            logger.info(f"Hydro normalization mode {normalization}")
-        else:
-            logger.info("No hydro normalization")
-
-            if "normalization" in resource:
-                resource.pop("normalization")
+        # get normalization before executing runoff
+        normalization = None
+        if ("normalization" in config) and isinstance(config["normalization"], dict):
+            normalization = config.pop("normalization")
 
         # check if there are hydro powerplants
         if resource["plants"].empty:
@@ -465,6 +583,54 @@ if __name__ == "__main__":
 
             if "clip_min_inflow" in config:
                 inflow = inflow.where(inflow > config["clip_min_inflow"], 0)
+
+            # check if normalization field belongs to the settings and it is not false
+            if normalization:
+                method, norm_year = normalization["method"], normalization["year"]
+                if method == "hydro_capacities":
+                    path_hydro_capacities = snakemake.input.hydro_capacities
+                    normalize_using_yearly = get_hydro_capacity_annual_hydro_generation(
+                        path_hydro_capacities, countries, norm_year
+                    ) * config.get("multiplier", 1.0)
+                elif method == "eia":
+                    path_eia_stats = snakemake.input.eia_hydro_generation
+                    normalize_using_yearly = get_eia_annual_hydro_generation(
+                        path_eia_stats, countries
+                    ) * config.get("multiplier", 1.0)
+
+                inflow = normalize_hydro(
+                    resource["plants"], inflow, normalize_using_yearly, norm_year
+                )
+                logger.info(
+                    f"Hydro normalization method '{method}' on year-statistics {norm_year}"
+                )
+            else:
+                logger.info("No hydro normalization")
+
+            inflow *= config.get("multiplier", 1.0)
+
+            # add zero values for out of hydrobasins elements
+            if len(bus_notin_hydrobasins) > 0:
+                regions_notin = regions.loc[
+                    bus_notin_hydrobasins, ["x", "y", "country"]
+                ]
+                logger.warning(
+                    f"Buses {bus_notin_hydrobasins} are not contained into hydrobasins."
+                    f"Setting empty time series. Bus location:\n{regions_notin}"
+                )
+
+                # initialize empty DataArray and append to inflow
+                notin_data = xr.DataArray(
+                    np.zeros(
+                        (len(bus_notin_hydrobasins), inflow.coords["time"].shape[0])
+                    ),
+                    dims=("plant", "time"),
+                    coords=dict(
+                        plant=bus_notin_hydrobasins,
+                        time=inflow.coords["time"],
+                    ),
+                )
+                inflow = xr.concat([inflow, notin_data], dim="plant")
 
             inflow.rename("inflow").to_netcdf(snakemake.output.profile)
     else: